moved AngularQuadratureGenerator to /quadrature in preparation for renaming QuadratureGenerator
