+++ src/quadrature/angular/tests/angular_quadrature_generator_mock.h
@@ -1,7 +1,7 @@
 #ifndef BART_SRC_QUADRATURE_ANGULAR_TESTS_ANGULAR_QUADRATURE_GENERATOR_MOCK_H_
 #define BART_SRC_QUADRATURE_ANGULAR_TESTS_ANGULAR_QUADRATURE_GENERATOR_MOCK_H_
 
-#include "quadrature/angular/angular_quadrature_generator_i.h"
+#include "quadrature/angular_quadrature_generator_i.h"
 
 #include "test_helpers/gmock_wrapper.h"
