+++ src/quadrature/angular_quadrature_generator_i.h
